+++ src/main/java/com/eCommerce/domain/Order.java
@@ -1,18 +1,19 @@
 package com.eCommerce.domain;
 
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Table;
+
 import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
 
 @Entity
 @Table(name="user_order")
@@ -26,8 +27,8 @@ public class Order {
 	private String shippingMethod;
 	private String orderStatus;
 	private BigDecimal orderTotal;
-	
-	@OneToMany(mappedBy = "order", cascade=CascadeType.ALL )
+
+	@OneToMany(mappedBy = "order", cascade = CascadeType.ALL)
 	private List<CartItem> cartItemList;
 	
 	@OneToOne(cascade=CascadeType.ALL)
