+++ src/main/java/com/eCommerce/domain/UserBilling.java
@@ -1,11 +1,11 @@
 package com.eCommerce.domain;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToOne;
 
 @Entity
 public class UserBilling {
@@ -20,8 +20,8 @@ public class UserBilling {
 	private String userBillingState;
 	private String userBillingCountry;
 	private String userBillingZipcode;
-	
-	@OneToOne(cascade=CascadeType.ALL)
+
+	@OneToOne(cascade = CascadeType.ALL)
 	private UserPayment userPayment;
 
 	public Long getId() {
