+++ src/main/java/com/eCommerce/service/UserPaymentService.java
@@ -4,6 +4,6 @@ import com.eCommerce.domain.UserPayment;
 
 public interface UserPaymentService {
 	UserPayment findById(Long id);
-	
-	void deteteById(Long id);
+
+	void deleteById(Long id);
 }
