+++ src/main/resources/application.properties
@@ -5,7 +5,7 @@ spring.thymeleaf.cache=false
 # ===============================
 
 # Set here configurations for the database connection
-spring.datasource.url=jdbc:mysql://localhost:3306/ecommercedatabase
+spring.datasource.url=jdbc:mysql://localhost:3306/ecommerceDB?createDatabaseIfNotExist=true&useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=Asia/Jakarta&useSSL=false&allowPublicKeyRetrieval=true
 
 # Username and secret
 spring.datasource.username=root
@@ -32,7 +32,7 @@ spring.jpa.show-sql=true
 spring.jpa.hibernate.ddl-auto = update
 
 # Allows Hibernate to generate SQL optimized for a particular DBMS
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 
 
 spring.mail.host=smtp.gmail.com
